requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
easyocr==1.7.1
Pillow==10.4.0
jinja2==3.1.4
